fix(radiometry): stop one stray pixel from breaking float input matching
A float raster could yield no key points at all. The float range was never the
problem: measured on a Landsat-8 / Sentinel-2 crop, native uint16 and the same
data rescaled to [0,1] both returned 3983 key points. Adding a single outlier
pixel, or a single +Inf, returned none.

The cause is that the stretch to 8 bit took its bounds from the array's minimum
and maximum, so the two most extreme pixels decided the mapping for the whole
scene and everything else collapsed into a handful of DN values. np.nanmax also
carries +Inf straight through. Float products routinely hold such pixels where
integer sensor data does not, which is why this looked like a float bug.

Floats now take their bounds from the 2nd and 98th percentile of the finite
pixels. Every failing case recovers: outlier, +Inf and a -9999 fill block all
return ~3900 key points instead of zero. Integer rasters keep the previous
minimum/maximum stretch, which leaves every existing result and the committed
end-to-end reference data byte-identical; the percentile costs about 2% of key
points on integer data and that switch deserves its own reviewed change.

Two hypotheses were measured and rejected rather than implemented: normalising
the reference and monitored images jointly was worse (3893 against 3908), and
restricting the percentile to the valid mask made no difference.

The same stretch existed in four diverging copies. They now share
core.radiometry.to_uint8, which is global_align's already-correct percentile
version. For coarse_to_fine this is a fix too, not just deduplication, since
nanpercentile drops NaN but keeps infinities; it keeps its own tighter bounds,
which is what the four-scene validation was run with.

Also: --no-value accepts floats, as a float raster's fill value need not be a
whole number; the low dynamic range warning no longer fires for every float
image, having compared a [0,1] span against a threshold of 10; and the README no
longer advises multiplying float data by 100, which is measurably a no-op.

+++ karios/cli/commands.py
@@ -207,7 +207,7 @@ def cli() -> None:
 )
 @click.option(
     "--no-value",
-    type=int,
+    type=float,
     multiple=True,
     default=None,
     help="Filter out key points where reference or monitored image has this DN value. Can be used multiple times (e.g., --no-value 0 --no-value 255)",
@@ -252,7 +252,7 @@ def process(
     generate_intermediate_product: bool,
     generate_kp_chips: bool,
     vector_mask: Optional[Path],
-    no_value: tuple[int, ...],
+    no_value: tuple[float, ...],
     title_prefix: Optional[str],
     dem_description: Optional[str],
     enable_large_shift_detection: bool,
